Fix 2-5: change 'int' to 'float'

+++ 2/5.py
@@ -1,5 +1,5 @@
 def c2f(cel):
   return cel * 1.8 + 32.0
 
-cel = int(input('Please enter a Celsius value: '))
+cel = float(input('Please enter a Celsius value: '))
 print(cel, 'degrees Celsius is', c2f(cel), 'degrees Fahrenheit.')
